requirements.txt
Django==2.1.7
django-allauth==0.38.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
psycopg2==2.7.7
pytz==2018.9
flake8==3.7.6
isort==4.3.4
model-mommy==1.6.0
pyhunter==0.3
